+++ src/coldStorage/ColdStorage.ts
@@ -19,6 +19,7 @@ import GuardTurn from '../utils/GuardTurn';
 import GuardPkHandler from '../handlers/GuardPkHandler';
 import DatabaseHandler from '../db/DatabaseHandler';
 import { DefaultLoggerFactory } from '@rosen-bridge/abstract-logger';
+import { TokenHandler } from '../handlers/tokenHandler';
 
 const logger = DefaultLoggerFactory.getInstance().getLogger(import.meta.url);
 
@@ -57,9 +58,16 @@ class ColdStorage {
         return;
       }
 
+      const thresholds = Configs.thresholds()[chainName].tokens;
+      if (Object.keys(thresholds).length === 0) {
+        logger.info(
+          `Abort cold storage process on chain [${chainName}] since no thresholds is set for it`
+        );
+        return;
+      }
+
       const chain = ChainHandler.getInstance().getChain(chainName);
       const lockedAssets = await chain.getLockAddressAssets();
-      const thresholds = Configs.thresholds()[chainName].tokens;
 
       let transferringNativeToken = 0n;
       const transferringTokens: TokenInfo[] = [];
@@ -73,9 +81,9 @@ class ColdStorage {
           return;
         }
         const isNativeToken =
-          Configs.tokenMap.search(chainName, {
-            [Configs.tokenMap.getIdKey(chainName)]: tokenId,
-          })[0][chainName].metaData.type === 'native';
+          TokenHandler.getInstance().getTokenMap().search(chainName, {
+            tokenId,
+          })[0][chainName].type === 'native';
         if (isNativeToken) {
           if (lockedAssets.nativeToken > thresholds[tokenId].high)
             transferringNativeToken =