+++ src/synchronization/EventSynchronization.ts
@@ -1,4 +1,4 @@
-import { Communicator, ECDSA, GuardDetection } from '@rosen-bridge/tss';
+import { GuardDetection } from '@rosen-bridge/detection';
 import { Semaphore } from 'await-semaphore';
 import { DefaultLoggerFactory } from '@rosen-bridge/abstract-logger';
 import {
@@ -28,6 +28,7 @@ import MinimumFeeHandler from '../handlers/MinimumFeeHandler';
 import ChainHandler from '../handlers/ChainHandler';
 import EventOrder from '../event/EventOrder';
 import DetectionHandler from '../handlers/DetectionHandler';
+import { Communicator } from '@rosen-bridge/communication';
 
 const logger = DefaultLoggerFactory.getInstance().getLogger(import.meta.url);
 
@@ -46,7 +47,7 @@ class EventSynchronization extends Communicator {
   protected constructor(detection: GuardDetection) {
     super(
       logger,
-      new ECDSA(Configs.tssKeys.secret),
+      Configs.tssKeys.encryptor,
       EventSynchronization.sendMessageWrapper,
       Configs.tssKeys.pubs.map((pub) => pub.curvePub),
       GuardTurn.UP_TIME_LENGTH
@@ -64,7 +65,7 @@ class EventSynchronization extends Communicator {
    * initializes EventSynchronization
    */
   static init = async () => {
-    const detection = DetectionHandler.getInstance().getDetection().curve;
+    const detection = DetectionHandler.getInstance().getDetection();
     // TODO: Definition of the required guard in the detection is not in the duty of this module!
     //  local:ergo/rosen-bridge/guard-service#428
     detection.setNeedGuardThreshold(GuardPkHandler.getInstance().requiredSign);
@@ -206,9 +207,8 @@ class EventSynchronization extends Communicator {
   protected getPeerIdByIndex = async (
     index: number
   ): Promise<string | undefined> => {
-    const pk = this.guardPks[index];
     const activeGuards = await this.detection.activeGuards();
-    return activeGuards.find((_) => _.publicKey === pk)?.peerId;
+    return activeGuards.find((_) => _.index === index)?.peerId;
   };
 
   /**